requirements.txt
streamlit
google-generativeai
google-cloud-vision
faiss
vertexai
langchain
langchain_google_genai
langchain-community